task6

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -4,7 +4,7 @@
 from typing import List, Union 
 
 
-def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_list: List[Union[int, float]]) -> float:
     """return the sum of float and int  mxd_list"""
     mxdlistsum = 0
     for i in mxd_list:
